requirements.txt
apache-airflow-providers-openlineage>=1.8.0
apache-airflow-providers-telegram
apache-airflow-providers-mysql
pandas
sqlalchemy
mysql-connector-python
python-dotenv
pytest
great-expectations